rename collector by validator
