+++ src/Notification/Validator/ValidatorInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Pledg\Bundle\PaymentBundle\Notification\Collector;
+namespace Pledg\Bundle\PaymentBundle\Notification\Validator;
 
 interface ValidatorInterface
 {
