+++ src/Resources/config/services.yml
@@ -11,7 +11,7 @@ services:
     Pledg\Bundle\PaymentBundle\Method\Config\Provider\PledgConfigProviderInterface: '@Pledg\Bundle\PaymentBundle\Method\Config\Provider\PledgConfigProvider'
     Pledg\Bundle\PaymentBundle\Method\View\Factory\PledgViewFactoryInterface: '@Pledg\Bundle\PaymentBundle\Method\View\Factory\PledgViewFactory'
     Pledg\Bundle\PaymentBundle\Method\Factory\PledgPaymentMethodFactoryInterface: '@Pledg\Bundle\PaymentBundle\Method\Factory\PledgPaymentMethodFactory'
-    Pledg\Bundle\PaymentBundle\Notification\Collector\ValidatorInterface: '@Pledg\Bundle\PaymentBundle\Notification\Collector\StandardValidator'
+    Pledg\Bundle\PaymentBundle\Notification\Validator\ValidatorInterface: '@Pledg\Bundle\PaymentBundle\Notification\Validator\StandardValidator'
     Pledg\Bundle\PaymentBundle\Provider\OrderProviderInterface: '@Pledg\Bundle\PaymentBundle\Provider\OrderProvider'
     Pledg\Bundle\PaymentBundle\Provider\PaymentTransactionProviderInterface: '@Pledg\Bundle\PaymentBundle\Provider\PaymentTransactionProvider'
     Pledg\Bundle\PaymentBundle\RedirectUrl\ParametersFactoryInterface: '@Pledg\Bundle\PaymentBundle\RedirectUrl\ParametersFactory'
@@ -69,7 +69,7 @@ services:
 
     Pledg\Bundle\PaymentBundle\Method\Factory\PledgPaymentMethodFactory: ~
 
-    Pledg\Bundle\PaymentBundle\Notification\Collector\StandardValidator: ~
+    Pledg\Bundle\PaymentBundle\Notification\Validator\StandardValidator: ~
 
     Pledg\Bundle\PaymentBundle\Method\Provider\PledgMethodProvider:
         tags:
