+++ tests/Unit/EventListener/Callback/PledgListenerBuilder.php
@@ -7,8 +7,8 @@ namespace Tests\Pledg\Bundle\PaymentBundle\Unit\EventListener\Callback;
 use Oro\Bundle\PaymentBundle\Method\Provider\PaymentMethodProviderInterface;
 use Pledg\Bundle\PaymentBundle\EventListener\Callback\PledgListener;
 use Pledg\Bundle\PaymentBundle\Method\PledgMethod;
-use Pledg\Bundle\PaymentBundle\Notification\Collector\StandardValidator;
-use Pledg\Bundle\PaymentBundle\Notification\Collector\ValidatorInterface;
+use Pledg\Bundle\PaymentBundle\Notification\Validator\StandardValidator;
+use Pledg\Bundle\PaymentBundle\Notification\Validator\ValidatorInterface;
 use Prophecy\Argument;
 use Prophecy\Prophet;
 use Symfony\Component\HttpFoundation\RequestStack;
